requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
qdrant-client
aiohttp
requests
streamlit